% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.R
\name{print.panel_causal_result}
\alias{print.panel_causal_result}
\title{Print a panel_causal_result}
\usage{
\method{print}{panel_causal_result}(x, ...)
}
\arguments{
\item{x}{an object of class panel_causal_result.}

\item{...}{unused, for compatibility.}
}
\description{
Print a panel_causal_result
}
